implement ceiling estimation for Pereira2018-linear
