+++ brainscore_language/artificial_subject.py
@@ -99,7 +99,12 @@ class ArtificialSubject:
 
     class RecordingType:
         """ method of recording """
-        spikerate_exact = "spikerate_exact"  # the exact spike-rate activity of each neuron
+
+        spikerate_exact = "spikerate_exact"
+        """ the exact spike-rate activity of each neuron """
+
+        fMRI = "fMRI"
+        """ functional magnetic resonance imaging """
 
     def perform_neural_recording(self, recording_target: RecordingTarget, recording_type: RecordingType):
         """
