+++ brainscore_language/plugins/pereira2018/test.py
@@ -1,4 +1,7 @@
+import copy
+
 import numpy as np
+import pytest
 from numpy.random import RandomState
 from pytest import approx
 
@@ -34,34 +37,50 @@ class TestBenchmark:
             self.neural_activity = neural_activity
 
         def digest_text(self, stimuli):
-            np.testing.assert_array_equal(stimuli, self.neural_activity['stimulus'])
+            assert len(stimuli) == len(self.neural_activity['presentation'])
+            self.neural_activity['stimulus'] = 'presentation', stimuli  # copy over
             return {'neural': self.neural_activity}
 
         def perform_neural_recording(self, recording_target: ArtificialSubject.RecordingTarget,
                                      recording_type: ArtificialSubject.RecordingType):
             assert recording_target == ArtificialSubject.RecordingTarget.language_system
-            assert recording_type == ArtificialSubject.RecordingType.spikerate_exact
+            assert recording_type == ArtificialSubject.RecordingType.fMRI
 
-    def test_dummy_bad(self):
-        benchmark = load_benchmark('Pereira2018-linear')
-        neural_activity = RandomState(0).random(size=(30, 25))  # todo
+    @pytest.mark.parametrize('experiment, expected_score', [
+        (243, approx(0.0017534 / .35378928, abs=0.001)),
+        (384, approx(0.01215216 / .36343748, abs=0.001)),
+    ])
+    def test_dummy_bad(self, experiment, expected_score):
+        benchmark = load_benchmark(f'Pereira2018.{experiment}-linear')
+        neural_activity = RandomState(0).random(size=(experiment, 25))  # presentation x neuroid
         neural_activity = NeuroidAssembly(neural_activity,
-                                          coords={'stimulus_id': ('presentation', np.arange(30)),
-                                                  'stimulus_category': ('presentation', ['a', 'b', 'c'] * 10),
+                                          coords={'stimulus_seq': ('presentation', np.arange(experiment)),
+                                                  'stimulus_num': ('presentation', np.arange(experiment)),
                                                   'neuroid_id': ('neuroid', np.arange(25)),
                                                   'region': ('neuroid', ['some_region'] * 25)},
                                           dims=['presentation', 'neuroid'])
         dummy_model = TestBenchmark.DummyModel(neural_activity=neural_activity)
         score = benchmark(dummy_model)
-        assert score == approx(0.0098731 / .318567, abs=0.001)
+        assert score == expected_score
 
-    def test_exact(self):
-        benchmark = load_benchmark('Pereira2018-linear')
-        dummy_model = TestBenchmark.DummyModel(neural_activity=benchmark.data)
+    @pytest.mark.parametrize('experiment', [
+        243,
+        384,
+    ])
+    def test_exact(self, experiment):
+        benchmark = load_benchmark(f'Pereira2018.{experiment}-linear')
+        exact_data = copy.deepcopy(benchmark.data).reset_index('presentation')
+        del exact_data['stimulus_id'], exact_data['stimulus']
+        exact_data = exact_data.set_index(presentation=list(exact_data['presentation'].coords))
+        dummy_model = TestBenchmark.DummyModel(neural_activity=exact_data)
         score = benchmark(dummy_model)
         assert score == approx(1)
 
-    def test_ceiling(self):
-        benchmark = load_benchmark('Pereira2018-linear')
+    @pytest.mark.parametrize('experiment, expected_ceiling', [
+        (243, .35378928),
+        (384, .36343748),
+    ])
+    def test_ceiling(self, experiment, expected_ceiling):
+        benchmark = load_benchmark(f'Pereira2018.{experiment}-linear')
         ceiling = benchmark.ceiling
-        assert ceiling == approx(.318567, abs=.0005)
+        assert ceiling == approx(expected_ceiling, abs=.0005)
