+++ brainscore_language/utils/__init__.py
@@ -1,0 +1,3 @@
+def fullname(obj):
+    """ Resolve the full module-qualified name of an object. Typically used for logger naming. """
+    return obj.__module__ + "." + obj.__class__.__name__
